+++ addon/synthDrivers/_inflectMicro/lib/onnxruntime/capi/build_and_package_info.py
@@ -1,2 +1,2 @@
-package_name = 'onnxruntime'
-__version__ = '1.28.0'
+package_name = 'onnxruntime-directml'
+__version__ = '1.24.4'