+++ addon/synthDrivers/_inflectMicro/lib/onnxruntime/quantization/operators/maxpool.py
@@ -1,5 +1,3 @@
-from ..quant_utils import FLOAT8_TYPES
-from .base_operator import QuantOperatorBase
 from .direct_q8 import Direct8BitOp, QDQDirect8BitOp
 
 
@@ -13,12 +11,8 @@ class QMaxPool(Direct8BitOp):
 
         # if version is less than 12, go to normal quantize.
         if self.quantizer.opset_version < 12:
-            QuantOperatorBase.quantize(self)
+            super(Direct8BitOp, self).quantize()
             return
-
-        # FP8 types are not supported for MaxPool; emit node unquantized.
-        if self.quantizer.activation_qType in FLOAT8_TYPES:
-            return QuantOperatorBase.quantize(self)
 
         # Direct 8bits op
         return super().quantize()
@@ -34,10 +28,6 @@ class QDQMaxPool(QDQDirect8BitOp):
 
         # if version is less than 12, just no change
         if self.quantizer.opset_version < 12:
-            return
-
-        # FP8 types are not supported for MaxPool; leave node unquantized.
-        if self.quantizer.activation_qType in FLOAT8_TYPES:
             return
 
         # Direct 8bits op