+++ addon/synthDrivers/_inflectMicro/lib/onnxruntime/tools/pytorch_export_helpers.py
@@ -1,8 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
-from __future__ import annotations
-
 import inspect
 from collections import abc