+++ addon/synthDrivers/_inflectMicro/lib/onnxruntime/transformers/onnx_model.py
@@ -2,7 +2,6 @@
 # Copyright (c) Microsoft Corporation.  All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from __future__ import annotations
 
 import itertools
 import logging
@@ -10,10 +9,6 @@ import os
 import sys
 from collections import deque
 from pathlib import Path
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from shape_infer_helper import SymbolicShapeInferenceHelper
 
 from float16 import convert_float_to_float16
 from onnx import (
@@ -28,6 +23,7 @@ from onnx import (
     save_model,
 )
 from onnx.external_data_helper import load_external_data_for_tensor, uses_external_data
+from shape_infer_helper import SymbolicShapeInferenceHelper
 
 logger = logging.getLogger(__name__)
 
@@ -55,8 +51,6 @@ class OnnxModel:
     def infer_runtime_shape(self, dynamic_axis_mapping={}, update=False):  # noqa: B006
         if self.enable_shape_infer:
             if self.shape_infer_helper is None or update:
-                from shape_infer_helper import SymbolicShapeInferenceHelper  # noqa: PLC0415
-
                 self.shape_infer_helper = SymbolicShapeInferenceHelper(self.model)
 
             try:
@@ -770,8 +764,6 @@ class OnnxModel:
         if use_symbolic_shape_infer:
             # Use symbolic shape inference since custom operators (like Gelu, SkipLayerNormalization etc)
             # are not recognized by onnx shape inference.
-            from shape_infer_helper import SymbolicShapeInferenceHelper  # noqa: PLC0415
-
             shape_infer_helper = SymbolicShapeInferenceHelper(model)
             try:
                 model_with_shape = shape_infer_helper.infer_shapes(model, auto_merge=True, guess_output_rank=False)